src/main/webapp/app/entities/delivery-man/update/delivery-man-update.component.ts
import { Component, OnInit } from '@angular/core';
import { HttpResponse } from '@angular/common/http';
import { ActivatedRoute } from '@angular/router';
import { Observable } from 'rxjs';
import { finalize } from 'rxjs/operators';

import { DeliveryManFormService, DeliveryManFormGroup } from './delivery-man-form.service';
import { IDeliveryMan } from '../delivery-man.model';
import { DeliveryManService } from '../service/delivery-man.service';

@Component({
  selector: 'jhi-delivery-man-update',
  templateUrl: './delivery-man-update.component.html',
})
export class DeliveryManUpdateComponent implements OnInit {
  isSaving = false;
  deliveryMan: IDeliveryMan | null = null;

  editForm: DeliveryManFormGroup = this.deliveryManFormService.createDeliveryManFormGroup();

  constructor(
    protected deliveryManService: DeliveryManService,
    protected deliveryManFormService: DeliveryManFormService,
    protected activatedRoute: ActivatedRoute
  ) {}

  ngOnInit(): void {
    this.activatedRoute.data.subscribe(({ deliveryMan }) => {
      this.deliveryMan = deliveryMan;
      if (deliveryMan) {
        this.updateForm(deliveryMan);
      }
    });
  }

  previousState(): void {
    window.history.back();
  }

  save(): void {
    this.isSaving = true;
    const deliveryMan = this.deliveryManFormService.getDeliveryMan(this.editForm);
    if (deliveryMan.id !== null) {
      this.subscribeToSaveResponse(this.deliveryManService.update(deliveryMan));
    } else {
      this.subscribeToSaveResponse(this.deliveryManService.create(deliveryMan));
    }
  }

  protected subscribeToSaveResponse(result: Observable<HttpResponse<IDeliveryMan>>): void {
    result.pipe(finalize(() => this.onSaveFinalize())).subscribe({
      next: () => this.onSaveSuccess(),
      error: () => this.onSaveError(),
    });
  }

  protected onSaveSuccess(): void {
    this.previousState();
  }

  protected onSaveError(): void {
    // Api for inheritance.
  }

  protected onSaveFinalize(): void {
    this.isSaving = false;
  }

  protected updateForm(deliveryMan: IDeliveryMan): void {
    this.deliveryMan = deliveryMan;
    this.deliveryManFormService.resetForm(this.editForm, deliveryMan);
  }
}
